+++ akshare/__init__.py
@@ -2229,9 +2229,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.8.10 fix: fix stock_a_all_pb interface
 1.8.11 fix: fix macro_china_ppi interface
 1.8.12 fix: fix stock_yjyg_em interface
+1.8.13 fix: fix macro_china_new_house_price interface
 """
 
-__version__ = "1.8.12"
+__version__ = "1.8.13"
 __author__ = "AKFamily"
 
 import sys