+++ docs/changelog.md
@@ -42,6 +42,12 @@
 
 ## 更新说明
 
+1.8.13 fix: fix macro_china_new_house_price interface
+
+    1. 修复 macro_china_new_house_price 接口
+    2. 修复 macro_china_enterprise_boom_index 接口
+    3. 修复 macro_china_national_tax_receipts 接口
+
 1.8.12 fix: fix stock_yjyg_em interface
 
     1. 修复 stock_yjyg_em 接口
@@ -1575,6 +1581,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.13 fix: fix macro_china_new_house_price interface
 
 1.8.12 fix: fix stock_yjyg_em interface